Updated the schemas for user.js, date.js and contact.js Still need to work on the controllers
