+++ controllers/contactsController.js
@@ -1,29 +1,29 @@
-const { Contact } = require("../models/User");
+var db = require("../models");
 
 // Defining methods for ContactControllers
 module.exports = {
     findAll: function (req, res) {
-        Contact.find(req.query)
+        db.Contact.find(req.query)
             .then(dbContact => res.json(dbContact))
             .catch(err => res.status(422).json(err));
     },
     findById: function (req, res) {
-        Contact.findById(req.params.id)
+        db/Contact.findById(req.params.id)
             .then(dbContact => res.json(dbContact))
             .catch(err => res.status(422).json(err));
     },
     create: function (req, res) {
-        Contact.create(req.body)
+        db.Contact.create(req.body)
             .then(dbContact => res.json(dbContact))
             .catch(err => res.status(422).json(err));
     },
     update: function (req, res) {
-        Contact.findOneAndUpdate({ id: req.params.id }, req, body)
+        db.Contact.findOneAndUpdate({ id: req.params.id }, req, body)
             .then(dbContact => res.json(dbContact))
             .catch(err => res.status(422).json(err));
     },
     remove: function (req, res) {
-        Contact.findById(req.params.id)
+        db.Contact.findById(req.params.id)
             .then(dbContact => dbContact.remove())
             .then(dbContact => res.json(dbContact))
             .catch(err => res.status(422).json(err));
