+++ models/Date.js
@@ -1,25 +1,30 @@
-// const mongoose = require("mongoose");
-// const Schema = mongoose.Schema;
-// // Create Schema for a new date
+const mongoose = require("mongoose");
+const Schema = mongoose.Schema;
 
-// const DatesSchema = new Schema({
-//     dateName: {
-//       type: String,
-//       required: false
-//     },
-//     dateNumber: {
-//       type: String,
-//       required: false
-//     },
-//     metThrough: {
-//       type: String,
-//       required: false
-//     },
-//     dateDescription: {
-//       type: String,
-//       required: false
-//     }
-//   });
+// Create Schema for a new date
 
-// const Dates = mongoose.model("Date", DatesSchema);
-// module.exports = Dates;
+const DatesSchema = new Schema({
+    _nameId: {
+        type: Schema.Types.ObjectId,
+        ref: "User"
+    },
+    dateName: {
+        type: String,
+        required: false
+    },
+    dateNumber: {
+        type: String,
+        required: false
+    },
+    metThrough: {
+        type: String,
+        required: false
+    },
+    dateDescription: {
+        type: String,
+        required: false
+    }
+});
+
+const Dates = mongoose.model("Date", DatesSchema);
+module.exports = Dates;
