+++ models/User.js
@@ -1,56 +1,12 @@
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
-// const DatesSchema = require("./Date");
 
-// const ContactsSchema = require("./Contact");
-
-// Create Schema for a new user, date and contact
-
-
-const DatesSchema = new Schema({
-  dateName: {
-    type: String,
-    required: false
-  },
-  dateNumber: {
-    type: String,
-    required: false
-  },
-  metThrough: {
-    type: String,
-    required: false
-  },
-  dateDescription: {
-    type: String,
-    required: false
-  },
-  dateAddress: {
-    type: String,
-    required: false
-  }
-});
-
-const ContactsSchema = new Schema({
-  contactName: {
-    type: String,
-    required: false
-  },
-  contactNumber: {
-    type: String,
-    required: false
-  },
-  relationship: {
-    type: String,
-    required: false
-  }
-});
-
-
-// Create Schemas for user, who they are going on a date with and emergency contacts
+// Create Schemas for a new user
 const UserSchema = new Schema({
   name: {
     type: String,
-    required: true
+    required: true,
+    unique: { index: { unique: true } }
   },
   email: {
     type: String,
@@ -76,12 +32,7 @@ const UserSchema = new Schema({
     type: Date,
     default: Date.now
   },
-  dates: [DatesSchema],
-  contacts: [ContactsSchema]
 });
 
-
 const User = mongoose.model("User", UserSchema);
-const Contact = mongoose.model("Contact", ContactsSchema);
-const Dates = mongoose.model("Dates", DatesSchema);
 module.exports = User
