+++ models/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    User: require("./User"),
+    Date: require("./Date"),
+    Contact: require("./Contact")
+  };
